Refactor functions and separate out into diff files
